change the detekt rule to be more generalised
